Fix django request KeyError exception
The bug here is that META['CONTENT_TYPE'] does not always exist but the
current code assumes it does. This is partly to be blamed on django's
mod_python connector which always sets this variable.

So without this patch, mod_wsgi support failed at this exact spot since
it - rightfully - does not always define CONTENT_TYPE.