Revert unreleased commit : "Reduce useless imports in urls.py"
Do not remove the 'import *', this is what allows this module to provide
'handler404' and 'handler500'. We provide django's default handlers. In a
nutshell, this module inherits django.conf.urls.defaults and redefines
whatever it wants. We could import the needed bit explicitely, but if newer
django versions add other handlers, we'll get bit in the a** again. Django's
API is retarded, but whatever.

See http://docs.djangoproject.com/en/dev/topics/http/views/#the-404-page-not-found-view

This reverts commit 06c44447c8ab7013649e68b287779ce1a04c8b99.