Sanitize authentication and middleware handling, refs: #3868
Authentication is not always required especially when there is no database
around for handling users and sesssions. Make it optional defaulted to on and
handled by sjconf.

Things could be a lot cleaner if we had support for handling a sorted list of
optional authentication and middleware backends in say [webengine:middleware]
and [webengine:auth] configuration sections.